Basic single oscillator synth built
This works well, but will run into issues as soon as any state (including UseAppSelector) is brought into the component, because of React's rendering behaviour. It needs to be refactored so that all of the state exists within a nested component, and the top-level component simply triggers note-on and note-off events.